Move builder from from settingsConsole.js to utils
