+++ lib/modules/filteReddit/LineFilter.js
@@ -3,6 +3,7 @@
 import _ from 'lodash';
 import {
 	asyncFilter,
+	caseBuilder,
 	string,
 	waitForEvent,
 	downcast,
@@ -11,7 +12,6 @@ import {
 import * as CommentNavigator from '../commentNavigator';
 import * as Hover from '../hover';
 import * as FilteReddit from '../filteReddit';
-import * as SettingsConsole from '../settingsConsole';
 import * as Cases from './cases';
 import * as Modules from './../../core/modules';
 import { Filter } from './Filter';
@@ -102,10 +102,10 @@ export class LineFilter extends Filter {
 
 		if (!this.initialConditions) this.initialConditions = lastConditions;
 
-		const $builder = SettingsConsole.drawBuilderBlock(lastConditions, builderCases);
+		const $builder = caseBuilder.drawBuilderBlock(lastConditions, builderCases);
 
 		if (this.BaseCase.variant === 'ondemand' || this.BaseCase === Cases.Group) {
-			SettingsConsole.makeBuilderBlockSortable($builder);
+			caseBuilder.makeBuilderBlockSortable($builder);
 		} else {
 			// Remove controls if case isn't a group
 			$builder.find('.builderControls, .addBuilderBlock').remove();
@@ -113,7 +113,7 @@ export class LineFilter extends Filter {
 
 		const $builderBlock = $builder.find('> .builderBlock');
 		$builderBlock.on('change input', frameThrottle(() => {
-			const conditions = SettingsConsole.readBuilderBlock($builderBlock, builderCases);
+			const conditions = caseBuilder.readBuilderBlock($builderBlock, builderCases);
 			if (!_.isEqual(lastConditions, conditions)) {
 				lastConditions = conditions;
 				this.update(undefined, conditions);
