+++ lib/utils/index.js
@@ -187,4 +187,6 @@ export * as CreateElement from './createElement';
 
 export * as Alert from './alert';
 
+export * as caseBuilder from './caseBuilder';
+
 export * as string from './string';
